+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added "bypass" parameter to Rcc HSE configuration (breaking change)
 - Add "usbsrc" function to Rcc configuration, used for selecting USB clock source
 - For STM32F030, require use more specific feature flag, e.g. "stm32f030xc"
+- Add `embedded-hal` `blocking::i2c::Read` implementation for I2C
 
 ### Fixed
 - RCC: Correct code to enable PLL.